backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
httpx>=0.27.0
python-multipart>=0.0.9
mcp>=0.1.0
strands-agents>=0.1.0
